+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Changed
 
+-   [all] changed rollup config to reduce bundle sizes
 -   [all] updated `package.json` in all packages to mark them as
     free of side effects
 -   [core] replaced a few functions from lodash by own implementations